+++ multiboot2/Changelog.md
@@ -1,6 +1,8 @@
 # CHANGELOG for crate `multiboot2`
 ## Unreleased
-- **Breaking** Make functions of `InformationBuilder` chainable. They now consume the builder.
+- **BREAKING** Make functions of `InformationBuilder` chainable. They now consume the builder.
+- **BREAKING** Allow non-standard memory area types by using new pair of
+  corresponding types: `MemoryAreaTypeId` and `MemoryAreaType`.
 
 ## 0.16.0 (2023-06-23)
 - **BREAKING** renamed `MULTIBOOT2_BOOTLOADER_MAGIC` to `MAGIC`